feat(refactor): refactored dir name and contract structure next is tests
