+++ CommitPoolContract/.editorconfig
